\name{bayeslm-package}
\alias{bayeslm-package}
\docType{package}
\title{
Efficient sampling for Gaussian linear regression with arbitrary priors
}
\description{
  The elliptical slice sampler for Bayesian linear regression with shrinakge priors such as horseshoe, Laplace prior, ridge prior. 
}
\author{
P. Richard Hahn, Jingyu He and Hedibert Lopes

Maintainer: Jingyu He \email{jingyuhe@cityu.edu.hk}
}
\references{
Hahn, P. Richard, Jingyu He, and Hedibert Lopes. \emph{Efficient sampling for Gaussian linear regression with arbitrary priors} (2017).
}
\keyword{ package }
\seealso{
  bayeslm
}
